been away for a while and had to do a lot on the go and saved it locally:
added a checkout page with logic!

edited the layout

added functions for the loanserviceoimplementations

added the user profile for the admin and the student to check which item they have loaned

added a returned boolean to loanItem.

added alerts to the basket view in case the item is not available for the date selected

added a query for the filtering and checking of the items are availble for the given period of time (a simple check if the total stock during that time would be 0 )

Still lots to do !